ref(dynamic-sampling): Introduce new design
