+++ static/app/views/settings/project/filtersAndSampling/modal/handleXhrErrorResponse.tsx
